t5541: add tests for rejecting a push with a HTTP error
A Git server or the HTTP server fronting it might want to reject a push
by returning a HTTP error response based on its configuration. Such a
response allows it to respond before the client has finished sending the
POST body, which is particularly relevant when rejecting due to size but
would also be relevant for server-side errors.

Currently remote-curl ignores anything with a status code above 300 but
this means that we cannot pass along any information in these
situations. But we can let remote-curl pass along the response as usual
if the server gives us the right content-type.

In preparation for lifting this restriction, we prepare a set of tests
that cover error responses when they come at different phases of the
request.

Additionally remote-curl ignores any output from send-pack in the error
case, which leads to it not reporting anything to the caller, resulting
in an erroneous "Everything up-to-date".

We include a test with a generic error response as well for when we do
still want to throw away the contents of the error response from the
server.

Signed-off-by: Carlos Martín Nieto <cmn@dwim.me>