remote-curl: accept pkt-line content for errors too
A remote might want to signal an error with a non-2xx status code but
still provide a response for the client.

This can happen if a push is cut short by reaching the limit set by
receive.maxinputsize where the server would want to respond with an HTTP
error so it's allowed to cut the upload short, but it still wants to
pass along the response from git-receive-pack.

Instead of throwing away all response bodies with a status larger than
2xx, accept those when the server has set a git content-type indicating
that it does want to send data to the git client.

Signed-off-by: Carlos Martín Nieto <cmn@dwim.me>