send-pack: read gently when trying to read the status
In a case where the remote returns an error via e.g. a POST size limit
enforced by a web server that returns a generic error without a Git
content-type, we will not receive any status information, but we still
want to pass along a report to our caller so it can show a reasonable
error to the user.

Here, if we find EOF when trying to read the result from the remote, we
would die with an error message, and the user might be left with no push
status report or with a misleading one.

Instead, let's try to read gently and the rest of the logic can already
take care of sending a report to our caller indicating the remote said
nothing.

Signed-off-by: Carlos Martín Nieto <cmn@dwim.me>